Diff/requirements.txt
accelerate
torch 
torchvision 
torchaudio
transformers
diffusers
opencv-python
espnet
espnet-model-zoo
espnet-tts-frontend
pyopenjtalk